Fix a Free Parameter To Zero And Fit a Model Again

It fixes a designated free parameter in a lavaan object to zero and refit the model.

Usage

fix_to_zero(fit, par_id, store_fit = FALSE)

Arguments

fit

A lavaan-class object.

par_id

An integer. The row number of the free parameter in the parameter table of fit to be fixed.

store_fit

Logical. If TRUE, fit will be stored in the output. Default is FALSE.

Value

A fix_to_zero-class object, which is a list with these elements:call

fit0 is the lavaan output of the refitted object. NA if the fit failed for some reasons. A model must converged, passed lavaan's post check, variance covariance matrix of estimates can be computed, and has one more degree of freedom than the original model to be considered acceptable.

fit1 is the original lavaan output if store_fit is TRUE. It is NULL otherwise

par_id is the row number of the designated free parameter.

call is the original call to this function.

ptable0 is the parameter table of with the designated parameter fixed to zero. It can be used for diagnostic purpose if the fit failed.

fit0_error is the error message in refitting, if any. If no error, it is `NA``.

vcov_ok is TRUE if the variance covariance matrix of the estimates can be computed without error nor warning. FALSE otherwise.

vcov_msg is the message generated when using lavaan::lavInspect() to get the variance-covariance matrix of the parameter estimates of the refitted model. If TRUE, then no error nor warning. Can be used for diagnostic purposes.

converged: Whether the modified model converged.

post_check_passed: Whether the solution of the modified model passed lavaans post check.

fit_not_ok: If the fit failed for some reasons, the fit object, if available, is stored in this element rather than in fit0. such that the fit object can be retrieved for diagnostic purposed if necessary.

df_diff_one: Whether the difference in model degrees of freedom between the modified model and the original model is one.

Details

It modifies the parameter table of a lavaan-class object and then fits the model again.

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
data(data_sem16)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
"
fit <- sem(mod, data_sem16)
# Fix the factor covariance to zero
out <- fix_to_zero(fit, par_id = 15)
summary(out$fit0)
#> lavaan 0.6.17 ended normally after 49 iterations
#> 
#>   Estimator                                         ML
#>   Optimization method                           NLMINB
#>   Number of model parameters                        12
#> 
#>   Number of observations                           336
#> 
#> Model Test User Model:
#>                                                       
#>   Test statistic                                27.655
#>   Degrees of freedom                                 9
#>   P-value (Chi-square)                           0.001
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Latent Variables:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1 =~                                               
#>     x1                1.000                           
#>     x2                0.911    0.253    3.604    0.000
#>     x3                1.051    0.323    3.257    0.001
#>   f2 =~                                               
#>     x4                1.000                           
#>     x5                2.538    1.424    1.782    0.075
#>     x6                5.526    6.331    0.873    0.383
#> 
#> Covariances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>   f1 ~~                                               
#>     f2                0.000                           
#> 
#> Variances:
#>                    Estimate  Std.Err  z-value  P(>|z|)
#>    .x1                1.010    0.119    8.457    0.000
#>    .x2                1.067    0.111    9.573    0.000
#>    .x3                0.700    0.114    6.159    0.000
#>    .x4                1.198    0.096   12.500    0.000
#>    .x5                1.037    0.182    5.705    0.000
#>    .x6                0.515    0.775    0.664    0.507
#>     f1                0.286    0.110    2.595    0.009
#>     f2                0.022    0.031    0.723    0.469
#>