Likelihood Ratio Test P-Values

Compute the likelihood ratio test p-values for free parameters in a lavaan output.

Usage

lrtp(
  fit,
  op = c("~", "~~"),
  no_variances = TRUE,
  no_error_variances = TRUE,
  no_error_covariances = TRUE,
  ...
)

Arguments

fit

A lavaan-class object.

op

A character vector of lavaan model syntax operators. Free parameters of these operators will be included, unless excluded by other arguments. Default is c("~", "~~").

no_variances

Logical. If TRUE, the default, then all free variances are excluded. (Error variances are handled by no_error_variances.)

no_error_variances

Logical, If TRUE, the default, then all free error variances are excluded.

no_error_covariances

Logical. If TRUE, the default, then all free error covariances are excluded.

...

Optional arguments to be passed to lavaan::parameterEstimates().

Value

An lrt-class object, which is a data-frame-like object with the output of lavaan::parameterEstimates(), with a column LRTp for the LRT p-values. Has a print method print.lrtp().

Details

It find free parameters in a lavaan-class object, compute the likelihood ratio test (LRT) p-value for each of them when fixed to zero, and return a parameter estimates data frame with the LRT p-values included.

See also

Author

Shu Fai Cheung https://orcid.org/0000-0002-9871-9448

Examples

library(lavaan)
data(data_sem16)
mod <-
"
f1 =~ x1 + x2 + x3
f2 =~ x4 + x5 + x6
f3 =~ x7 + x8 + x9
f4 =~ x10 + x11 + x12
f2 ~~ f1
f3 ~ f1 + f2
f4 ~ f3
"
fit <- sem(mod, data_sem16)
lrtp(fit)
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err    Chisq     LRTp
#>   f3 ~                                                
#>     f1                0.248    0.097   10.343    0.001
#>     f2                0.292    0.199    3.778    0.052
#>   f4 ~                                                
#>     f3                2.118    0.684   86.205    0.000
#> 
#> Covariances:
#>                    Estimate  Std.Err    Chisq     LRTp
#>   f1 ~~                                               
#>     f2                0.039    0.022    9.717    0.002
#> 
lrtp(fit, op = "~")
#> 
#> Parameter Estimates:
#> 
#>   Standard errors                             Standard
#>   Information                                 Expected
#>   Information saturated (h1) model          Structured
#> 
#> Regressions:
#>                    Estimate  Std.Err    Chisq     LRTp
#>   f3 ~                                                
#>     f1                0.248    0.097   10.343    0.001
#>     f2                0.292    0.199    3.778    0.052
#>   f4 ~                                                
#>     f3                2.118    0.684   86.205    0.000
#>